Moved the RecipeManager related classes to their own Package.
