+++ core/akka-runtime/src/main/scala/com/ing/baker/runtime/recipe_manager/PollingAware.scala
@@ -1,9 +1,7 @@
-package com.ing.baker.runtime
+package com.ing.baker.runtime.recipe_manager
 
-import com.ing.baker.il.CompiledRecipe
 import com.ing.baker.runtime.common.RecipeRecord
 import com.typesafe.scalalogging.LazyLogging
-import shapeless.Lazy
 
 import scala.concurrent.{ExecutionContext, Future}
